Split all this up.
