+++ tests/integration/Context/CannotEditFilesUserContext.php
@@ -1,0 +1,28 @@
+<?php
+/**
+ * User context where user cannot edit files.
+ *
+ * @package wp-crontrol
+ */
+
+namespace Crontrol\Tests\Context;
+
+/**
+ * User context where user cannot edit files (no edit_files capability).
+ */
+class CannotEditFilesUserContext extends TestUserContext {
+	#[\Override]
+	public function can_create_php_cron_events(): bool {
+		return false;
+	}
+
+	#[\Override]
+	public function can_edit_php_cron_events(): bool {
+		return false;
+	}
+
+	#[\Override]
+	public function can_delete_php_cron_events(): bool {
+		return false;
+	}
+}
